google switched to a HTTPS repo for new installs
fedora-workstation-repos suggests this happened by 2022.

I'd stumbling over this once or twice, and ignored it.  But this last time,
the problem became clear.

I think existing installs are staying on HTTP for now.  But I don't want to
assume the HTTP repo will stay around forever!

Also, for some reason, new installs *using this role* switch the APT repo
to HTTPS after the initial install.  That's just bad.  Let's make sure we
fail *fast*, if HTTPS is going to cause a failure.

Document the problem this causes with apt-cacher-ng, in our README.